Moved results bar to top, made it so everything under the faction selector scrolls when overflowing
