+++ src/pages/ti4-combat-calculator/Ti4CombatCalculator.page.css
@@ -1,23 +1,17 @@
 .ti4-combat-calc {
+    --ti4-combat-calc__padding: 20px;
+
     display: grid;
     grid-template-columns: 1fr 1fr;
-    grid-template-rows: auto;
+    grid-template-rows: 40px 1fr;
     grid-template-areas: 
-        "header header"
-        "attacker defender"
-        "simbutton simbutton"
-        "results results";
-    row-gap: 20px;
-    column-gap: 20px;
-    margin: 20px;
-}
-
-.ti4-combat-calc__header {
-    grid-area: header;
-}
-
-.ti4-combat-calc__simbutton {
-    grid-area: simbutton;
+        "results results"
+        "attacker defender";
+    row-gap: var(--ti4-combat-calc__padding);
+    column-gap: var(--ti4-combat-calc__padding);
+    height: calc(100vh - (var(--ti4-combat-calc__padding) * 2));
+    width: calc(100vw - (var(--ti4-combat-calc__padding) * 2));
+    padding: var(--ti4-combat-calc__padding);
 }
 
 .ti4-combat-calc__attacker {
@@ -30,7 +24,7 @@
 
 .ti-combat-calc__participant {
     border: 1px gray solid;
-    padding: 20px;
+    overflow-y: hidden;
 }
 
 .ti4-combat-calc__results {
