+++ src/pages/ti4-combat-calculator/components/FactionFleetBuilderForm/FactionFleetBuilderForm.component.css
@@ -1,8 +1,17 @@
 .faction-fleet-builder-form {
     display: grid;
     grid-template-columns: 1fr;
-    grid-template-rows: auto;
+    grid-template-rows: auto 1fr;
     row-gap: 10px;
     column-gap: 10px;
-    margin: 10px;
+    height: 100%;
+}
+
+.faction-fleet-builder-form__faction-select-container {
+    padding: 20px 20px 0px 20px;
+}
+
+.faction-fleet-builder-form__fleet-builder-form {
+    overflow-y: auto;
+    padding: 20px;
 }
