+++ src/pages/ti4-combat-calculator/components/FactionFleetBuilderForm/components/FleetBuilderForm/FleetBuilderForm.component.css
@@ -1,7 +1,7 @@
 .fleet-builder-form {
   display: grid;
   grid-template-columns: 1fr;
-  grid-template-rows: auto;
+  grid-template-rows: min-content;
   row-gap: 20px;
 }
